move parse_age_group to utils
