+++ leap/data_generation/occurrence_data.py
@@ -7,6 +7,7 @@ import plotly.express as px
 import statsmodels.api as sm
 import statsmodels.formula.api as smf
 from leap.utils import get_data_path, poly
+from leap.data_generation.utils import parse_age_group
 from leap.logger import get_logger
 from typing import Tuple, Dict, Any
 from statsmodels.genmod.generalized_linear_model import GLMResultsWrapper
@@ -18,36 +19,6 @@ logger = get_logger(__name__, 20)
 STARTING_YEAR = 2000
 MAX_YEAR = 2019
 MAX_AGE = 65
-
-
-def parse_age_group(x: str, max_age: int = MAX_AGE) -> Tuple[int, int]:
-    """Parse an age group string into a tuple of integers.
-    
-    Args:
-        x: The age group string. Must be in the format "X-Y", "X+", "X-Y years", "<1 year".
-
-    Returns:
-        A tuple of integers representing the lower and upper age of the age group.
-
-    Examples:
-    
-        >>> parse_age_group("0-4")
-        (0, 4)
-        >>> parse_age_group("5-9 years")
-        (5, 9)
-        >>> parse_age_group("10+")
-        (10, 65)
-        >>> parse_age_group("<1 year")
-        (0, 1)
-    """
-    if x == "<1 year":
-        return 0, 1
-    elif "-" in x:
-        return int(x.split(" ")[0].split("-")[0]), int(x.split(" ")[0].split("-")[1])
-    elif "+" in x:
-        return int(x.split("+")[0]), max_age
-    else:
-        raise ValueError(f"Invalid age group: {x}")
 
 
 def load_asthma_df(starting_year: int = STARTING_YEAR) -> pd.DataFrame:
@@ -85,7 +56,7 @@ def load_asthma_df(starting_year: int = STARTING_YEAR) -> pd.DataFrame:
     # Age groups are in the format "X-Y" or "80+"
     # Set the age to the average of the age group
     df["age"] = df.apply(
-        lambda x: int(np.mean(parse_age_group(x["age_group"]))), # type: ignore
+        lambda x: int(np.mean(parse_age_group(x["age_group"], max_age=MAX_AGE))), # type: ignore
         axis=1
     ) # type: ignore
